Move some stuff about, such as StringRect to its own header.
