+++ Code/GraphMol/MolDraw2D/DrawShape.cpp
@@ -11,8 +11,9 @@
 //
 
 #include <GraphMol/MolDraw2D/DrawShape.h>
+#include <GraphMol/MolDraw2D/DrawText.h>
 #include <GraphMol/MolDraw2D/MolDraw2D.h>
-#include <GraphMol/MolDraw2D/MolDraw2DDetails.h>
+#include <GraphMol/MolDraw2D/StringRect.h>
 
 namespace RDKit {
 namespace MolDraw2D_detail {
