+++ Code/GraphMol/MolDraw2D/MolDraw2DDetails.cpp
@@ -173,93 +173,6 @@ std::vector<Point2D> getBracketPoints(
   return {p0, p1, p2, p3};
 }
 
-namespace {
-void drawArrow(MolDraw2D &drawer, const MolDrawShape &shape) {
-  PRECONDITION(shape.shapeType == MolDrawShapeType::Arrow, "bad shape type");
-  PRECONDITION(shape.points.size() == 4, "bad points size");
-  drawer.setColour(shape.lineColour);
-  drawer.setLineWidth(shape.lineWidth);
-  drawer.drawLine(shape.points[0], shape.points[1]);
-  if (!shape.fill) {
-    drawer.drawLine(shape.points[1], shape.points[2]);
-    drawer.drawLine(shape.points[1], shape.points[3]);
-  } else {
-    drawer.setFillPolys(true);
-    std::vector<Point2D> head(shape.points.begin() + 1, shape.points.end());
-    drawer.drawPolygon(head);
-  }
-}
-void drawPolyline(MolDraw2D &drawer, const MolDrawShape &shape) {
-  PRECONDITION(shape.shapeType == MolDrawShapeType::Polyline, "bad shape type");
-  PRECONDITION(shape.points.size() > 1, "not enough points");
-  drawer.setColour(shape.lineColour);
-  auto lw = shape.lineWidth;
-  if (shape.scaleLineWidth) {
-    lw *= drawer.scale() * 0.02;
-    if (lw < 0.0) {
-      lw = 0.0;
-    }
-  }
-  drawer.setLineWidth(lw);
-  if (shape.points.size() > 2 && shape.fill) {
-    drawer.setFillPolys(true);
-  } else {
-    drawer.setFillPolys(false);
-  }
-  if (drawer.drawOptions().comicMode) {
-    auto drawPoints =
-        handdrawnLine(shape.points[0], shape.points[1], drawer.scale());
-    for (unsigned int i = 2; i < shape.points.size(); ++i) {
-      auto lpts = MolDraw2D_detail::handdrawnLine(
-          shape.points[i - 1], shape.points[i], drawer.scale());
-      std::move(lpts.begin(), lpts.end(), std::back_inserter(drawPoints));
-    }
-    drawer.drawPolygon(drawPoints);
-  } else {
-    if (shape.points.size() > 2) {
-      drawer.drawPolygon(shape.points);
-    } else {
-      drawer.drawLine(shape.points[0], shape.points[1]);
-    }
-  }
-}
-void drawEllipse(MolDraw2D &drawer, const MolDrawShape &shape) {
-  PRECONDITION(shape.shapeType == MolDrawShapeType::Ellipse, "bad shape type");
-  PRECONDITION(shape.points.size() == 2, "wrong points");
-  drawer.setColour(shape.lineColour);
-  drawer.setLineWidth(shape.lineWidth);
-  if (shape.fill) {
-    drawer.setFillPolys(true);
-  } else {
-    drawer.setFillPolys(false);
-  }
-  drawer.drawEllipse(shape.points[0], shape.points[1]);
-}
-}  // namespace
-void drawShapes(MolDraw2D &drawer, const std::vector<MolDrawShape> &shapes) {
-  const auto ocolour = drawer.colour();
-  const auto olw = drawer.lineWidth();
-  const auto ofill = drawer.fillPolys();
-  for (const auto &shape : shapes) {
-    switch (shape.shapeType) {
-      case MolDrawShapeType::Polyline:
-        drawPolyline(drawer, shape);
-        break;
-      case MolDrawShapeType::Arrow:
-        drawArrow(drawer, shape);
-        break;
-      case MolDrawShapeType::Ellipse:
-        drawEllipse(drawer, shape);
-        break;
-      default:
-        ASSERT_INVARIANT(false, "unrecognized shape type");
-    }
-  }
-  drawer.setColour(ocolour);
-  drawer.setLineWidth(olw);
-  drawer.setFillPolys(ofill);
-};
-
 // there are a several empirically determined constants here.
 std::vector<Point2D> handdrawnLine(Point2D cds1, Point2D cds2, double scale,
                                    bool shiftBegin, bool shiftEnd,
@@ -299,144 +212,6 @@ std::vector<Point2D> handdrawnLine(Point2D cds1, Point2D cds2, double scale,
   }
   pts.push_back(cds2);
   return pts;
-}
-
-// ****************************************************************************
-bool doesLineIntersect(const StringRect &rect, const Point2D &end1,
-                       const Point2D &end2, double padding) {
-  Point2D tl, tr, bl, br;
-  rect.calcCorners(tl, tr, br, bl, padding);
-  if (doLinesIntersect(end2, end1, tl, tr, nullptr)) {
-    return true;
-  }
-  if (doLinesIntersect(end2, end1, tr, br, nullptr)) {
-    return true;
-  }
-  if (doLinesIntersect(end2, end1, br, bl, nullptr)) {
-    return true;
-  }
-  if (doLinesIntersect(end2, end1, bl, tl, nullptr)) {
-    return true;
-  }
-  return false;
-}
-
-// ****************************************************************************
-bool doesTriangleIntersect(const StringRect &rect, const Point2D &pt1,
-                           const Point2D &pt2, const Point2D &pt3,
-                           double padding) {
-  if (doesLineIntersect(rect, pt1, pt2, padding)) {
-    return true;
-  }
-  if (doesLineIntersect(rect, pt2, pt3, padding)) {
-    return true;
-  }
-  if (doesLineIntersect(rect, pt3, pt1, padding)) {
-    return true;
-  }
-  return false;
-}
-
-// ****************************************************************************
-bool doesLineIntersectEllipse(const Point2D &centre, double xradius,
-                              double yradius, double padding,
-                              const Point2D &end1, const Point2D &end2) {
-  // using
-  // https://math.stackexchange.com/questions/76457/check-if-a-point-is-within-an-ellipse
-  // to see if either end1 or end2 are inside the ellipse.
-  double xr2 = (xradius + padding) * (xradius + padding);
-  double yr2 = (yradius + padding) * (yradius + padding);
-  double xdisc = (end1.x - centre.x) * (end1.x - centre.x) / xr2;
-  double ydisc = (end1.y - centre.y) * (end1.y - centre.y) / yr2;
-  if (xdisc + ydisc <= 1.0) {
-    return true;
-  }
-  xdisc = (end2.x - centre.x) * (end2.x - centre.x) / xr2;
-  ydisc = (end2.y - centre.y) * (end2.y - centre.y) / yr2;
-  if (xdisc + ydisc <= 1.0) {
-    return true;
-  }
-  return false;
-}
-
-// ****************************************************************************
-bool doesLineIntersectArc(const Point2D &centre, double xradius, double yradius,
-                          double start_ang, double stop_ang, double padding,
-                          const Point2D &end1, const Point2D &end2) {
-  double xr = xradius + padding;
-  double yr = yradius + padding;
-  double xr2 = xr * xr;
-  double yr2 = yr * yr;
-
-  auto pointInArc = [&](const Point2D &p) -> bool {
-    double xdisc = p.x * p.x / xr2;
-    double ydisc = p.y * p.y / yr2;
-    // start_ang can be more than stop_ang, if, for example, the arc goes from
-    // 315 to 45.
-    if (xdisc + ydisc <= 1.0) {
-      // end1 is inside the whole ellipse.  See if the angle it makes with the
-      // x axis lies between start_and and stop_ang.
-      double th = atan2(p.x, p.y) * 180.0 / M_PI;
-      if (th < 0.0) {
-        th += 360.0;
-      }
-      if (start_ang < stop_ang) {
-        // it's pretty straightforward
-        if (th >= start_ang && th <= stop_ang) {
-          return true;
-        }
-      } else {
-        // the arc crosses 0.
-        if (th >= start_ang && th <= 360.0) {
-          return true;
-        }
-        if (th >= 0.0 && th <= stop_ang) {
-          return true;
-        }
-      }
-    }
-    return false;
-  };
-
-  Point2D p1 = end1 - centre;
-  if (pointInArc(p1)) {
-    return true;
-  }
-  Point2D p2 = end2 - centre;
-  if (pointInArc(p2)) {
-    return true;
-  }
-  return false;
-}
-
-// ****************************************************************************
-bool doLinesIntersect(const Point2D &l1s, const Point2D &l1f,
-                      const Point2D &l2s, const Point2D &l2f, Point2D *ip) {
-  // using spell from answer 2 of
-  // https://stackoverflow.com/questions/563198/how-do-you-detect-where-two-line-segments-intersect
-  double s1_x = l1f.x - l1s.x;
-  double s1_y = l1f.y - l1s.y;
-  double s2_x = l2f.x - l2s.x;
-  double s2_y = l2f.y - l2s.y;
-
-  double d = (-s2_x * s1_y + s1_x * s2_y);
-  if (d == 0.0) {
-    // parallel lines.
-    return false;
-  }
-  double s, t;
-  s = (-s1_y * (l1s.x - l2s.x) + s1_x * (l1s.y - l2s.y)) / d;
-  t = (s2_x * (l1s.y - l2s.y) - s2_y * (l1s.x - l2s.x)) / d;
-
-  if (s >= 0 && s <= 1 && t >= 0 && t <= 1) {
-    if (ip) {
-      ip->x = l1s.x + t * s1_x;
-      ip->y = l1s.y + t * s1_y;
-    }
-    return true;
-  }
-
-  return false;
 }
 
 }  // namespace MolDraw2D_detail
